Keep injected Vert.x helper free of agent-only types
QueryExecutorUtil is injected into the application class loader, so it can only reference types that class loader can resolve. Narrow it to the package-private access it actually needs and move the VertxSqlClientData typing into VertxSqlClientUtil on the agent side.
