+++ instrumentation/vertx/vertx-sql-client/vertx-sql-client-5.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/vertx/sqlclient/v5_0/QueryExecutorInstrumentation.java
@@ -22,7 +22,6 @@ import io.opentelemetry.javaagent.instrumentation.vertx.sqlclient.common.v4_0.Ve
 import io.opentelemetry.javaagent.instrumentation.vertx.sqlclient.common.v4_0.VertxSqlClientUtil;
 import io.vertx.core.internal.PromiseInternal;
 import io.vertx.sqlclient.SqlConnectOptions;
-import io.vertx.sqlclient.impl.QueryExecutorUtil;
 import io.vertx.sqlclient.internal.PreparedStatement;
 import java.util.Collection;
 import javax.annotation.Nullable;
@@ -51,7 +50,7 @@ class QueryExecutorInstrumentation implements TypeInstrumentation {
     @Advice.OnMethodExit(suppress = Throwable.class, inline = false)
     public static void onExit(@Advice.This Object queryExecutor) {
       // copy client data from ThreadLocal to VirtualField
-      QueryExecutorUtil.setData(
+      VertxSqlClientUtil.setQueryExecutorData(
           queryExecutor, new VertxSqlClientData(getSqlConnectOptions(), getDbSystem()));
     }
   }
@@ -110,7 +109,7 @@ class QueryExecutorInstrumentation implements TypeInstrumentation {
           return new AdviceScope(callDepth);
         }
 
-        VertxSqlClientData data = QueryExecutorUtil.getData(queryExecutor);
+        VertxSqlClientData data = VertxSqlClientUtil.getQueryExecutorData(queryExecutor);
         if (data == null) {
           return new AdviceScope(callDepth);
         }
