+++ instrumentation/vertx/vertx-sql-client/vertx-sql-client-5.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/vertx/sqlclient/v5_0/SqlConnectionBaseInstrumentation.java
@@ -5,6 +5,7 @@
 
 package io.opentelemetry.javaagent.instrumentation.vertx.sqlclient.v5_0;
 
+import static io.opentelemetry.javaagent.instrumentation.vertx.sqlclient.common.v4_0.VertxSqlClientUtil.attachPreparedStatementData;
 import static io.opentelemetry.javaagent.instrumentation.vertx.sqlclient.common.v4_0.VertxSqlClientUtil.wrapContext;
 import static net.bytebuddy.matcher.ElementMatchers.named;
 import static net.bytebuddy.matcher.ElementMatchers.returns;
@@ -16,7 +17,6 @@ import io.opentelemetry.javaagent.instrumentation.vertx.sqlclient.common.v4_0.Ve
 import io.vertx.core.Future;
 import io.vertx.sqlclient.PreparedStatement;
 import io.vertx.sqlclient.SqlConnectOptions;
-import io.vertx.sqlclient.impl.QueryExecutorUtil;
 import io.vertx.sqlclient.internal.SqlClientBase;
 import io.vertx.sqlclient.internal.SqlConnectionBase;
 import javax.annotation.Nullable;
@@ -69,8 +69,7 @@ class SqlConnectionBaseInstrumentation implements TypeInstrumentation {
               ? VertxSqlClientSingletons.getConnectOptionsDbSystem(connectOptions)
               : null;
       return wrapContext(
-          QueryExecutorUtil.attachPreparedStatementData(
-              future, new VertxSqlClientData(connectOptions, dbSystem)));
+          attachPreparedStatementData(future, new VertxSqlClientData(connectOptions, dbSystem)));
     }
   }
 }
