+++ instrumentation/vertx/vertx-sql-client/vertx-sql-client-5.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/vertx/sqlclient/v5_0/VertxSqlClientInstrumentationModule.java
@@ -45,14 +45,6 @@ public class VertxSqlClientInstrumentationModule extends InstrumentationModule
   }
 
   @Override
-  public List<String> exposedClassNames() {
-    // QueryExecutorUtil is injected into the application class loader, it needs to be able to
-    // resolve the type that it stores in virtual fields
-    return singletonList(
-        "io.opentelemetry.javaagent.instrumentation.vertx.sqlclient.common.v4_0.VertxSqlClientData");
-  }
-
-  @Override
   public List<TypeInstrumentation> typeInstrumentations() {
     return asList(
         new CommandSchedulerInstrumentation(),
